backups/backup_20250321_102317/requirements.txt
streamlit==1.29.0
pandas==2.0.2
Pillow==9.5.0
matplotlib==3.7.1
numpy==1.24.3
python-dotenv==1.0.0
requests-oauthlib==1.3.1
pytest==7.4.3
pytest-mock==3.12.0
pytest-cov==4.1.0